+++ climpred/tests/test_hindcastEnsemble_class.py
@@ -16,37 +16,10 @@ def test_hindcastEnsemble_init_da(hind_da_initialized_1d):
     assert hindcast
 
 
-def test_add_reference(hind_ds_initialized_1d, reconstruction_ds_1d):
-    """Test to see if a reference can be added to the HindcastEnsemble"""
-    # TODO: This should be removed once `add_reference` is deprecated.
-    hindcast = HindcastEnsemble(hind_ds_initialized_1d)
-    hindcast = hindcast.add_reference(reconstruction_ds_1d, 'reconstruction')
-    # Will fail if this comes back empty.
-    assert hindcast.get_reference()
-
-
-def test_add_reference_da(hind_ds_initialized_1d, observations_da_1d):
-    """Test to see if a reference can be added to the HindcastEnsemble as a da"""
-    # TODO: This should be removed once `add_reference` is deprecated.
-    hindcast = HindcastEnsemble(hind_ds_initialized_1d)
-    hindcast = hindcast.add_reference(observations_da_1d, 'observations')
-    assert hindcast.get_reference()
-
-
-def test_add_reference_deprecated(hind_ds_initialized_1d, observations_da_1d):
-    """Tests that deprecation warning is thrown for add_reference method."""
-    hindcast = HindcastEnsemble(hind_ds_initialized_1d)
-    with pytest.warns(PendingDeprecationWarning) as record:
-        hindcast = hindcast.add_reference(observations_da_1d, 'observations')
-    assert 'deprecated' in record[0].message.args[0]
-
-
 def test_add_observations(hind_ds_initialized_1d, reconstruction_ds_1d):
     """Test to see if observations can be added to the HindcastEnsemble"""
-    # NOTE: This should be removed once `add_reference` is deprecated.
     hindcast = HindcastEnsemble(hind_ds_initialized_1d)
     hindcast = hindcast.add_observations(reconstruction_ds_1d, 'reconstruction')
-    # Will fail if this comes back empty.
     assert hindcast.get_observations()
 
 
@@ -71,41 +44,6 @@ def test_add_hist_da_uninitialized_1d(hind_ds_initialized_1d, hist_da_uninitiali
     assert hindcast.get_uninitialized()
 
 
-def test_compute_metric(
-    hind_ds_initialized_1d, reconstruction_ds_1d, observations_ds_1d
-):
-    """Test to see if compute_metric can be run from the HindcastEnsemble"""
-    # TODO: Remove this test after deprecating `compute_metric()`
-    hindcast = HindcastEnsemble(hind_ds_initialized_1d)
-    hindcast = hindcast.add_observations(reconstruction_ds_1d, 'reconstruction')
-    hindcast = hindcast.add_observations(observations_ds_1d, 'observations')
-    # Don't need to check for NaNs, etc. since that's handled in the prediction
-    # module testing.
-    hindcast.compute_metric()  # compute over all observations
-    # compute over single observation
-    hindcast.compute_metric('reconstruction')
-    # test all keywords
-    hindcast.compute_metric(metric='rmse', comparison='m2o')
-
-
-def test_compute_metric_single(hind_ds_initialized_1d, reconstruction_ds_1d):
-    """Test to see if compute_metric automatically works with a single observational
-    product."""
-    # TODO: Remove this test after deprecating `compute_metric()`
-    hindcast = HindcastEnsemble(hind_ds_initialized_1d)
-    hindcast = hindcast.add_observations(reconstruction_ds_1d, 'reconstruction')
-    hindcast.compute_metric()
-
-
-def test_compute_metric_deprecated(hind_ds_initialized_1d, reconstruction_ds_1d):
-    """Tests that deprecation warning is thrown for compute_metric method."""
-    hindcast = HindcastEnsemble(hind_ds_initialized_1d)
-    hindcast = hindcast.add_observations(reconstruction_ds_1d, 'reconstruction')
-    with pytest.warns(PendingDeprecationWarning) as record:
-        hindcast = hindcast.compute_metric()
-    assert 'deprecated' in record[0].message.args[0]
-
-
 @pytest.mark.slow
 def test_verify(hind_ds_initialized_1d, reconstruction_ds_1d, observations_ds_1d):
     """Test to see if verify can be run from the HindcastEnsemble"""
@@ -126,39 +64,6 @@ def test_verify_single(hind_ds_initialized_1d, reconstruction_ds_1d):
     hindcast = HindcastEnsemble(hind_ds_initialized_1d)
     hindcast = hindcast.add_observations(reconstruction_ds_1d, 'reconstruction')
     hindcast.verify()
-
-
-@pytest.mark.skip(reason='will be addressed when refactoring hindcast stuff.')
-def test_compute_uninitialized(
-    hind_ds_initialized_1d,
-    hist_ds_uninitialized_1d,
-    reconstruction_ds_1d,
-    observations_ds_1d,
-):
-    """Test to see if compute_uninitialized can be frun from the HindcastEnsemble"""
-    hindcast = HindcastEnsemble(hind_ds_initialized_1d)
-    hindcast = hindcast.add_observations(reconstruction_ds_1d, 'reconstruction')
-    hindcast = hindcast.add_uninitialized(hist_ds_uninitialized_1d)
-    # single observations, no declaration of name.
-    hindcast.compute_uninitialized()
-    hindcast = hindcast.add_observations(observations_ds_1d, 'observations')
-    # multiple observations, no name declaration.
-    hindcast.compute_uninitialized()
-    # multiple observations, call one.
-    hindcast.compute_uninitialized('reconstruction')
-    hindcast.compute_uninitialized(metric='rmse', comparison='m2o')
-
-
-def test_compute_persistence(
-    hind_ds_initialized_1d, reconstruction_ds_1d, observations_ds_1d
-):
-    """Test to see if compute_persistence can be run from the HindcastEnsemble"""
-    hindcast = HindcastEnsemble(hind_ds_initialized_1d)
-    hindcast = hindcast.add_observations(reconstruction_ds_1d, 'reconstruction')
-    hindcast = hindcast.add_observations(observations_ds_1d, 'observations')
-    hindcast.compute_persistence()
-    hindcast.compute_persistence('observations')
-    hindcast.compute_persistence(metric='rmse')
 
 
 def test_smooth_goddard(reconstruction_ds_3d, hind_ds_initialized_3d):
@@ -226,27 +131,6 @@ def test_get_uninitialized(hind_ds_initialized_1d, hist_ds_uninitialized_1d):
     hindcast = hindcast.add_uninitialized(hist_ds_uninitialized_1d)
     uninit = hindcast.get_uninitialized()
     assert uninit == hindcast._datasets['uninitialized']
-
-
-def test_get_reference(hind_ds_initialized_1d, reconstruction_ds_1d):
-    """Tests whether get_reference method works."""
-    hindcast = HindcastEnsemble(hind_ds_initialized_1d)
-    hindcast = hindcast.add_reference(reconstruction_ds_1d, 'FOSI')
-    # Without name keyword.
-    ref = hindcast.get_reference()
-    assert ref == hindcast._datasets['observations']['FOSI']
-    # With name keyword.
-    ref = hindcast.get_reference('FOSI')
-    assert ref == hindcast._datasets['observations']['FOSI']
-
-
-def test_get_reference_deprecated(hind_ds_initialized_1d, reconstruction_ds_1d):
-    """Tests that deprecation warning is thrown for get_reference method."""
-    hindcast = HindcastEnsemble(hind_ds_initialized_1d)
-    hindcast = hindcast.add_reference(reconstruction_ds_1d, 'FOSI')
-    with pytest.warns(PendingDeprecationWarning) as record:
-        hindcast = hindcast.get_reference()
-    assert 'deprecated' in record[0].message.args[0]
 
 
 def test_get_observations(hind_ds_initialized_1d, reconstruction_ds_1d):